wip: reimagining setup
