+++ OutboxKit.sln
@@ -11,10 +11,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{2C24
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DapperMySqlSample", "samples\DapperMySqlSample\DapperMySqlSample.csproj", "{C479A883-CE60-4521-A5D5-725C5704D736}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core.Polling", "src\Core.Polling\Core.Polling.csproj", "{10DF0B0F-8B21-4D3C-AE81-7EB23A9727C4}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core.Push", "src\Core.Push\Core.Push.csproj", "{7234FCF1-F0FD-4449-90E8-08E97879D505}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -32,19 +28,9 @@ Global
 		{C479A883-CE60-4521-A5D5-725C5704D736}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C479A883-CE60-4521-A5D5-725C5704D736}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C479A883-CE60-4521-A5D5-725C5704D736}.Release|Any CPU.Build.0 = Release|Any CPU
-		{10DF0B0F-8B21-4D3C-AE81-7EB23A9727C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{10DF0B0F-8B21-4D3C-AE81-7EB23A9727C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{10DF0B0F-8B21-4D3C-AE81-7EB23A9727C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{10DF0B0F-8B21-4D3C-AE81-7EB23A9727C4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7234FCF1-F0FD-4449-90E8-08E97879D505}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7234FCF1-F0FD-4449-90E8-08E97879D505}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7234FCF1-F0FD-4449-90E8-08E97879D505}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7234FCF1-F0FD-4449-90E8-08E97879D505}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{9E17490D-8407-441F-AF33-620EAD39C9AD} = {E1748C96-94CD-479F-BF71-585EB56C9FBF}
 		{C479A883-CE60-4521-A5D5-725C5704D736} = {2C240B41-C6C9-4921-9139-ECB2AA468316}
-		{10DF0B0F-8B21-4D3C-AE81-7EB23A9727C4} = {E1748C96-94CD-479F-BF71-585EB56C9FBF}
-		{7234FCF1-F0FD-4449-90E8-08E97879D505} = {E1748C96-94CD-479F-BF71-585EB56C9FBF}
 	EndGlobalSection
 EndGlobal
