+++ samples/DapperMySqlSample/Program.cs
@@ -1,15 +1,13 @@
-using System.Data;
-using Dapper;
-using MySqlConnector;
+using DapperMySqlSample;
 using YakShaveFx.OutboxKit.Core;
+using YakShaveFx.OutboxKit.Core.Polling;
 
 var builder = WebApplication.CreateBuilder(args);
 
 builder.Services
     .AddSingleton<FakeTargetProducer>()
     .AddOutboxBatchFetcher<OutboxBatchFetcher>()
-    .AddOutboxKit(kit =>
-        kit.WithTargetProducer<FakeTargetProducer>("fake"));
+    .AddOutboxKit(kit => kit.WithTargetProducer<FakeTargetProducer>("fake"));
 
 var app = builder.Build();
