+++ src/Core/Observability.cs
@@ -1,11 +1,10 @@
 using System.Diagnostics;
-using System.Reflection;
 
 namespace YakShaveFx.OutboxKit.Core;
 
 internal static class Observability
 {
-    internal static string ActivitySourceName { get; } = typeof(Observability).Assembly.GetName().Name!;
+    private static string ActivitySourceName { get; } = typeof(Observability).Assembly.GetName().Name!;
 
     internal static readonly ActivitySource ActivitySource = new(
         ActivitySourceName,
