+++ src/Core/OutboxSettings.cs
@@ -1,14 +1,8 @@
-// --------------------------------------------------------------------------------------------------------------------
-// <copyright company="Openvia">
-//     Copyright (c) Openvia. All rights reserved.
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
+using YakShaveFx.OutboxKit.Core.Polling;
 
 namespace YakShaveFx.OutboxKit.Core;
 
 public sealed class OutboxSettings
 {
-    public int MaxBatchSize { get; init; } = 100;
-    
-    public TimeSpan PollingInterval { get; init; } = TimeSpan.FromMinutes(5);
+    public PollingSettings Polling { get; init; } = new();
 }
